packages/csc-gemini/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csc-gemini"
version = "0.1.0"
description = "Gemini AI IRC client for CSC system."
requires-python = ">=3.8"
dependencies = [
    "csc-ai-api>=0.1.0",
    "google-generativeai>=0.5.0",
]

[project.scripts]
csc-gemini = "csc_gemini.main:main"

[tool.setuptools.packages.find]
where = ["."]
